expression: add vector constructor
Motivation for this changes is to provide a distinguishable interface
for vector type expressions.

The square bracket literal is ambigious for lists and vectors,
so that we need to perform a distinction not using CQL layer.
At first we should use the collection constructor to manage
both lists and vectors (although a vector is not a collection).
Later during preparation of expressions we should be able to get
to know the exact type using given receiver (column specification).

Knowing the type of expression we may use their respective constructor
(in this case the vector constructor being introduced),
which would make the implementation more precise.
In particular, this separates the implementation from collections,
which could've been quite misleading and inelegant.

This commit introduces vector constructor, functions making use of it,
and it's necessary visitor appearances.

However vector constructor is not yet used anywhere,
the next commit should adjust collection constructor and make use
of the vector constructor and it's features.